Demo is now just JS, passing json config
